Creates helper functions
Created two helper functions. wordRange() finds the lower and upper word
length limits of the double pointer passed into the function. toLower()
simply checks if the parameter is an uppercase letter and turns it
lowercase, if the parameter passed into the function is not an uppercase
letter, the function will return the original value back without
modifications.