Creates hashmap
Created a functional method that takes in the list of words to be
searched in the word search puzzle and stores it in a triple pointer.
There could be potential problems I have yet to see, but currently the
program is not halting for any unsuspecting reason. Changes might need
to be made to improve upon this implementation.