Scraps hashmaps and completes horizontal searches
This commit gets rid of the hashmap idea and all helper functions that
were purely associated with it, seeing as it is more difficult to be done in C than anticipated. This commit completes the ability to search for words horizontally, meaning from left to right, and from right to left. Helper functions were added in order to get this done. strrev() takes in the address of a string and reverses the order of the characters. strcmpi() takes in the addresses of two strings and checks them to determine if they both contain the same characters in the same order by returning a 0 if they are not equivalent and a 1 if they are. This method is case insensitive (accounts for the upper and lower case and purposefully ignores it). clearStr() takes in a string and sets every character in that string to be NULL.