Alphabetically ordered words found
Added helper functions and modified small segments of code in the
searchPuzzle() function. Instead of immediately printing when a word is
found, it adds the words found into a list called wordsFound which over
time accumulates all the found words and prints them using the
printFound() function. strcpyi() is a function made to copy strings
from one destination to another, different from strcpy() from the
string.h module because it accounts for a string's length better.
alphaSort() is a method that sorts a list of words into alphabetical
order. printFound() is a method meant to simply print the words
in a list with the words "Word found: " before it.