Completes TPS, fib.s, and recursion.s
Answers all questions from the TPS activities.
Modifies fib.s to take user input so the program can be more flexible in terms of what number
it's trying to find the fibonacci sequence to.
Edits recursion.s in order to mimic the instructions written in
recursion.c. recursion.s now takes in user input and uses recursion to
return the correct number.