Fixes minor calculation flaws
BobCatCandy-Alex_Woersching.s now accounts for the left over wrappers
that were not able to be traded in each transaction. This was also done
in bcc.c. An example of this can be seen in tests.md as Test 6.
BobCatCandy-Alex_Woersching.s now accounts for invalid inputs for the
price and wrapper exchange number by terminating the program when the
invalid input is entered.
Modifies the amount that's added to the stack pointer when newBars is
called. Modifies it from -12 to -8.
Removed all test.txt files and moved their contents into tests.md while
adding two more test cases.
Renamed BobCatCandy.s to BobCatCandy-Alex_Woersching.s.